Fix probable bug
Possible integer overflow when converting int -> uint16. This bug may never occur, but it is better to have proper handling of this case.

+++ pkg/reconciler/topology/partitionset/partitionset_controller_test.go
@@ -314,7 +314,7 @@ func TestReconcile(t *testing.T) {
 				require.Error(t, err, "expected an error")
 			} else {
 				require.NoError(t, err, "expected no error")
-				require.Equal(t, uint16(tc.wantPartitionCount), partitionSet.Status.Count)
+				require.Equal(t, uint(tc.wantPartitionCount), partitionSet.Status.Count)
 			}
 
 			if tc.wantPartitionsNotReady {
